+++ src/effects/builtin/CMakeLists.txt
@@ -31,6 +31,11 @@ set(MODULE_SRC
     ${CMAKE_CURRENT_LIST_DIR}/tonegen/chirpeffect.h
     ${CMAKE_CURRENT_LIST_DIR}/tonegen/toneeffect.cpp
     ${CMAKE_CURRENT_LIST_DIR}/tonegen/toneeffect.h
+
+    ${CMAKE_CURRENT_LIST_DIR}/reverb/reverbeffect.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/reverb/reverbeffect.h
+    ${CMAKE_CURRENT_LIST_DIR}/reverb/reverbviewmodel.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/reverb/reverbviewmodel.h
     )
 
 # AU3
