+++ src/effects/builtin/effects_builtin.qrc
@@ -5,5 +5,6 @@
         <file>tonegen/ChirpView.qml</file>
         <file>tonegen/ToneView.qml</file>
         <file>general/GeneralEffectView.qml</file>
+        <file>reverb/ReverbView.qml</file>
     </qresource>
 </RCC>
