+++ src/effects/builtin/internal/builtineffectsrepository.h
@@ -3,10 +3,14 @@
 */
 #pragma once
 
+#include <map>
+
 #include "modularity/ioc.h"
 #include "effects/effects_base/ieffectsviewregister.h"
 
 #include "../ibuiltineffectsrepository.h"
+
+#include "libraries/lib-components/ComponentInterfaceSymbol.h"
 
 namespace au::effects {
 class BuiltinEffectsRepository : public IBuiltinEffectsRepository
@@ -19,5 +23,10 @@ public:
     void init();
 
     EffectMetaList effectMetaList() const override;
+
+    EffectMeta effectMeta(const ComponentInterfaceSymbol& symbol) const;
+
+private:
+    std::map<ComponentInterfaceSymbol, EffectMeta> m_metas;
 };
 }
