+++ src/effects/builtin/reverb/reverbviewmodel.cpp
@@ -1,0 +1,7 @@
+#include "reverbviewmodel.h"
+
+using namespace au::effects;
+
+ReverbViewModel::ReverbViewModel()
+{
+}
