+++ src/effects/builtin/reverb/reverbviewmodel.h
@@ -1,0 +1,9 @@
+#pragma once
+
+namespace au::effects {
+class ReverbViewModel
+{
+public:
+    ReverbViewModel();
+};
+}
